Refactor directory structure and some names
