+++ frontEnd/src/APIs/Root/sendGetRequest.ts
@@ -1,9 +1,9 @@
-import serverIP from '../../assets/serverIP.json'
+import authServerIP from '../IPs/authServerIP.json'
 
 import withAutoResend, { RequestType } from './requestWrapperAPI'
 
 const sendGetRequest = async (data: any) => {
-    const response = await withAutoResend(RequestType.GET, `${serverIP}/restrictedAPI`)
+    const response = await withAutoResend(RequestType.GET, `${authServerIP}/restrictedAPI`)
 
     if (!response)
       return
