+++ frontEnd/src/components/helpers/PasswordStrengthChecker.tsx
